moved base environment var functions to core
