+++ test/conftest.py
@@ -9,7 +9,7 @@ from pathlib import Path
 import pytest
 
 from h2integrate import EXAMPLE_DIR
-from h2integrate.resource.utilities.env_var_api_tools import set_nlr_key_dot_env
+from h2integrate.resource.utilities.nlr_developer_api_keys import set_nlr_key_dot_env
 
 
 def pytest_sessionstart(session):
